refactor(server): isolate local to hosted authority transfer
